+++ resources/js/components/ict_request_divisi1/Verifikasi_from_email.vue
@@ -1,4 +1,20 @@
 <template>
+ <div class="p-grid crud-demo">
+    <div class="p-col-12">
+      <div class="card">
+        <Toast />
+        <ConfirmDialog> </ConfirmDialog>
+        <DataTable
+          :rows="25"
+          :loading="loading"
+        >
+          <template #loading>
+            Memverifikasi Link
+          </template>
+        </DataTable>   
+      </div>
+    </div>
+  </div>
 </template>
 <script>
 import moment from 'moment';
@@ -7,7 +23,8 @@ export default {
     return {
         verif:[],
         ireq_id:null,
-        todayyy : null
+        todayyy : null,
+        loading : true
     };
   },
   mounted() {
@@ -18,9 +35,20 @@ export default {
       localStorage.setItem("loggedIn", "true");
         this.axios.get('/api/cek-verif-id/'+this.$route.params.code).then((res)=>{
             this.verif = res.data;
+          if(res.data == null) {
+             this.$router.push({ name: 'Page Error', params: { stat: 'notvalid' } }) }
+          else{
             this.ireq_id = res.data.ireq_id;
-            this.todayyy = moment(new Date()).format('YYYY-MM-DD H:mm:s')
-            if(this.verif.expired_at <= this.todayyy){
+            this.todayyy = moment(new Date()).format('YYYY-MM-DD HH:mm:s')
+            console.log('today ',this.todayyy)
+            console.log('date expired ', this.verif.expired_at)
+
+            if (this.verif.expired_at >= this.todayyy){
+              this.loginUser();
+            }
+            
+            else if(this.verif.expired_at <= this.todayyy){
+             
               this.axios.get('/sanctum/csrf-cookie').then(() => {
               this.axios.post('/api/login-approval',this.verif).then((res)=>{
                 localStorage.clear();
@@ -28,13 +56,14 @@ export default {
                 localStorage.setItem("token", res.data.token);
                 localStorage.setItem("id", res.data.id);
                 localStorage.setItem("usr_name", res.data.usr_name);
-                this.$router.push('/page-error-410');
+                this.$router.push({ name: 'Page Error', params: { stat: 'expired' } }) 
+                this.loading = false;
+
+
               });
-            });
+             });
             }
-            if (this.verif.expired_at >= this.todayyy){
-              this.loginUser();
-            }
+           }
         });
       },
     loginUser(){
@@ -46,6 +75,8 @@ export default {
           localStorage.setItem("id", res.data.id);
           localStorage.setItem("usr_name", res.data.usr_name);
           this.$router.push({ name: 'Ict Request Verifikasi From Email', params: { code: this.ireq_id, status: this.$route.params.status } })
+          
+          this.loading = false;
         });
       });
     },
